+++ src/libfb/fb_private.h
@@ -81,58 +81,58 @@ struct fb_internal {
     uint32_t if_magic;
     uint32_t type_magic;
     /* Static information: per device TYPE.     */
-    int (*if_open)(struct fb_internal *ifp, const char *file, int _width, int _height);                       /**< @brief open device */
-    int (*if_open_existing)(struct fb_internal *ifp, int width, int height, struct fb_platform_specific *fb_p);                       /**< @brief open device */
+    int (*if_open)(struct fb_internal *ifp, const char *file, size_t _width, size_t _height);                       /**< @brief open device */
+    int (*if_open_existing)(struct fb_internal *ifp, size_t width, size_t height, struct fb_platform_specific *fb_p);                       /**< @brief open device */
     int (*if_close_existing)(struct fb_internal *ifp);                       /**< @brief close embedded fb */
     struct fb_platform_specific *(*if_existing_get)(uint32_t magic);                       /**< @brief allocate memory for platform specific container*/
     void                         (*if_existing_put)(struct fb_platform_specific *fb_p);                       /**< @brief free memory for platform specific container */
     int (*if_close)(struct fb_internal *ifp);                                                                 /**< @brief close device */
     int (*if_clear)(struct fb_internal *ifp, unsigned char *pp);                                              /**< @brief clear device */
-    ssize_t (*if_read)(struct fb_internal *ifp, int x, int y, unsigned char *pp, size_t count);               /**< @brief read pixels */
-    ssize_t (*if_write)(struct fb_internal *ifp, int x, int y, const unsigned char *pp, size_t count);        /**< @brief write pixels */
+    ssize_t (*if_read)(struct fb_internal *ifp, size_t x, size_t y, unsigned char *pp, size_t count);               /**< @brief read pixels */
+    ssize_t (*if_write)(struct fb_internal *ifp, size_t x, size_t y, const unsigned char *pp, size_t count);        /**< @brief write pixels */
     int (*if_rmap)(struct fb_internal *ifp, ColorMap *cmap);                                                  /**< @brief read colormap */
     int (*if_wmap)(struct fb_internal *ifp, const ColorMap *cmap);                                            /**< @brief write colormap */
-    int (*if_view)(struct fb_internal *ifp, int xcent, int ycent, int xzoom, int yzoom);                      /**< @brief set view */
-    int (*if_getview)(struct fb_internal *ifp, int *xcent, int *ycent, int *xzoom, int *yzoom);               /**< @brief get view */
-    int (*if_setcursor)(struct fb_internal *ifp, const unsigned char *bits, int xb, int yb, int xo, int yo);  /**< @brief define cursor */
-    int (*if_cursor)(struct fb_internal *ifp, int mode, int x, int y);                                        /**< @brief set cursor */
-    int (*if_getcursor)(struct fb_internal *ifp, int *mode, int *x, int *y);                                  /**< @brief get cursor */
-    int (*if_readrect)(struct fb_internal *ifp, int xmin, int ymin, int _width, int _height, unsigned char *pp);              /**< @brief read rectangle */
-    int (*if_writerect)(struct fb_internal *ifp, int xmin, int ymin, int _width, int _height, const unsigned char *pp);       /**< @brief write rectangle */
-    int (*if_bwreadrect)(struct fb_internal *ifp, int xmin, int ymin, int _width, int _height, unsigned char *pp);            /**< @brief read monochrome rectangle */
-    int (*if_bwwriterect)(struct fb_internal *ifp, int xmin, int ymin, int _width, int _height, const unsigned char *pp);     /**< @brief write rectangle */
-    int (*if_configure_window)(struct fb_internal *ifp, int width, int height);         /**< @brief configure window */
-    int (*if_refresh)(struct fb_internal *ifp, int x, int y, int w, int h);         /**< @brief refresh window */
+    int (*if_view)(struct fb_internal *ifp, size_t xcent, size_t ycent, size_t xzoom, size_t yzoom);                      /**< @brief set view */
+    int (*if_getview)(struct fb_internal *ifp, size_t *xcent, size_t *ycent, size_t *xzoom, size_t *yzoom);               /**< @brief get view */
+    int (*if_setcursor)(struct fb_internal *ifp, const unsigned char *bits, size_t xb, size_t yb, size_t xo, size_t yo);  /**< @brief define cursor */
+    int (*if_cursor)(struct fb_internal *ifp, int mode, size_t x, size_t y);                                        /**< @brief set cursor */
+    int (*if_getcursor)(struct fb_internal *ifp, int *mode, size_t *x, size_t *y);                                  /**< @brief get cursor */
+    int (*if_readrect)(struct fb_internal *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, unsigned char *pp);              /**< @brief read rectangle */
+    int (*if_writerect)(struct fb_internal *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, const unsigned char *pp);       /**< @brief write rectangle */
+    int (*if_bwreadrect)(struct fb_internal *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, unsigned char *pp);            /**< @brief read monochrome rectangle */
+    int (*if_bwwriterect)(struct fb_internal *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, const unsigned char *pp);     /**< @brief write rectangle */
+    int (*if_configure_window)(struct fb_internal *ifp, size_t width, size_t height);         /**< @brief configure window */
+    int (*if_refresh)(struct fb_internal *ifp, size_t x, size_t y, size_t w, size_t h);         /**< @brief refresh window */
     int (*if_poll)(struct fb_internal *ifp);          /**< @brief handle events */
     int (*if_flush)(struct fb_internal *ifp);         /**< @brief flush output */
     int (*if_free)(struct fb_internal *ifp);          /**< @brief free resources */
     int (*if_help)(struct fb_internal *ifp);          /**< @brief print useful info */
     char *if_type;      /**< @brief what "open" calls it */
-    int if_max_width;   /**< @brief max device width */
-    int if_max_height;  /**< @brief max device height */
+    size_t if_max_width;   /**< @brief max device width */
+    size_t if_max_height;  /**< @brief max device height */
     /* Dynamic information: per device INSTANCE. */
     char *if_name;      /**< @brief what the user called it */
-    int if_width;       /**< @brief current values */
-    int if_height;
+    size_t if_width;       /**< @brief current values */
+    size_t if_height;
     int if_selfd;       /**< @brief select(fd) for input events if >= 0 */
     /* Internal information: needed by the library.     */
     int if_fd;          /**< @brief internal file descriptor */
-    int if_xzoom;       /**< @brief zoom factors */
-    int if_yzoom;
-    int if_xcenter;     /**< @brief pan position */
-    int if_ycenter;
+    size_t if_xzoom;       /**< @brief zoom factors */
+    size_t if_yzoom;
+    size_t if_xcenter;     /**< @brief pan position */
+    size_t if_ycenter;
     int if_cursmode;    /**< @brief cursor on/off */
-    int if_xcurs;       /**< @brief cursor position */
-    int if_ycurs;
+    size_t if_xcurs;       /**< @brief cursor position */
+    size_t if_ycurs;
     unsigned char *if_pbase;/**< @brief Address of malloc()ed page buffer.      */
     unsigned char *if_pcurp;/**< @brief Current pointer into page buffer.       */
     unsigned char *if_pendp;/**< @brief End of page buffer.                     */
     int if_pno;         /**< @brief Current "page" in memory.           */
     int if_pdirty;      /**< @brief Page modified flag.                 */
-    long if_pixcur;     /**< @brief Current pixel number in framebuffer. */
-    long if_ppixels;    /**< @brief Sizeof page buffer (pixels).                */
+    size_t if_pixcur;     /**< @brief Current pixel number in framebuffer. */
+    size_t if_ppixels;    /**< @brief Sizeof page buffer (pixels).                */
     int if_debug;       /**< @brief Buffered IO debug flag.             */
-    long if_poll_refresh_rate; /**< @brief Recommended polling rate for interactive framebuffers in microseconds. */
+    size_t if_poll_refresh_rate; /**< @brief Recommended polling rate for interactive framebuffers in microseconds. */
     /* State variables for individual interface modules */
     union {
         char *p;
@@ -145,14 +145,14 @@ struct fb_internal {
  * Not for general consumption.
  */
 FB_EXPORT extern int _fb_disk_enable;
-FB_EXPORT extern int fb_sim_readrect(fb *ifp, int xmin, int ymin, int _width, int _height, unsigned char *pp);
-FB_EXPORT extern int fb_sim_writerect(fb *ifp, int xmin, int ymin, int _width, int _height, const unsigned char *pp);
-FB_EXPORT extern int fb_sim_bwreadrect(fb *ifp, int xmin, int ymin, int _width, int _height, unsigned char *pp);
-FB_EXPORT extern int fb_sim_bwwriterect(fb *ifp, int xmin, int ymin, int _width, int _height, const unsigned char *pp);
-FB_EXPORT extern int fb_sim_view(fb *ifp, int xcenter, int ycenter, int xzoom, int yzoom);
-FB_EXPORT extern int fb_sim_getview(fb *ifp, int *xcenter, int *ycenter, int *xzoom, int *yzoom);
-FB_EXPORT extern int fb_sim_cursor(fb *ifp, int mode, int x, int y);
-FB_EXPORT extern int fb_sim_getcursor(fb *ifp, int *mode, int *x, int *y);
+FB_EXPORT extern int fb_sim_readrect(fb *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, unsigned char *pp);
+FB_EXPORT extern int fb_sim_writerect(fb *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, const unsigned char *pp);
+FB_EXPORT extern int fb_sim_bwreadrect(fb *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, unsigned char *pp);
+FB_EXPORT extern int fb_sim_bwwriterect(fb *ifp, size_t xmin, size_t ymin, size_t _width, size_t _height, const unsigned char *pp);
+FB_EXPORT extern int fb_sim_view(fb *ifp, size_t xcenter, size_t ycenter, size_t xzoom, size_t yzoom);
+FB_EXPORT extern int fb_sim_getview(fb *ifp, size_t *xcenter, size_t *ycenter, size_t *xzoom, size_t *yzoom);
+FB_EXPORT extern int fb_sim_cursor(fb *ifp, int mode, size_t x, size_t y);
+FB_EXPORT extern int fb_sim_getcursor(fb *ifp, int *mode, size_t *x, size_t *y);
 
 __END_DECLS
